Fixed jerky RMB scrolling at large zoom factors, done by changing the grid fields 'orig_x' and 'orig_y' from int --> double.

+++ src/r_grid.h
@@ -38,8 +38,8 @@ public:
 	int mode;
 
 	// map coordinates for centre of canvas
-	int orig_x;
-	int orig_y;
+	double orig_x;
+	double orig_y;
 
 	// scale for drawing map
 	float Scale;
